added more options for search parameters
Added the option to search by year, category, or film title. 
Also added option to filter by winners

Half of the code is just to run it to make sure it works, the rest is the actual logic. The logic and control variables it uses will most likely be used in the final project and the rest discarded.